+++ gameEngine/actions.js
@@ -4,112 +4,30 @@
 /* eslint-disable no-plusplus */
 const { getRandomUpTo } = require('../gameTerminal/randomFunc');
 
-const {
-    // phase,
-    card: cardConst,
-    // target: targetConst,
-    action,
-} = require('../src/constants');
+const { card: cardConst, action } = require('../src/constants');
+
+const { waterCard } = require('./specials');
 
 const {
+    getActivePlayer,
+    getInActivePlayer,
+    setHealthToZero,
     removePanic,
-    waterCard,
-    turningHand,
-    showCards,
-    deleteCardsShown,
-    disableCards,
-} = require('./specials');
+    lastActionChange,
+    damagePlayer,
+    deleteCardsShownProperty,
+} = require('./helpers');
 
-
-function getActivePlayer(game) {
-    return game.players.find((p) => p.id === game.active);
-}
-
-function getInActivePlayer(game) {
-    return game.players.find((p) => p.id !== game.active);
-}
-
-// function to change lastAction game property (using for playing sound)
-function lastActionChange(game, lastAction) {
-    game.lastAction.type = lastAction;
-}
-
-// This function takes the player and the key for his card
-// and destination = from, then moves card to graveyard via deleting the key from array
-function moveCardGraveyard(player, key, from, opponent) {
-    if (from === cardConst.ITEMCARD) {
-    // console.log('moveCardGraveyard called for item ', player.hero, key);
-        player.grave[key] = player.item[key];
-        delete player.item[key];
-    } else if (from === cardConst.ITEMINACTIVE) {
-    // console.log('moveCardGraveyard called for itemOpponent ', opponent.hero, opponent.grave);
-        opponent.grave[key] = player.item[key];
-        delete player.item[key];
-    } else {
-    // console.log('moveCardGraveyard called ', player, key);
-        player.grave[key] = player.hand[key];
-        // we check if card has panic  == false property (forest Mushroom card)
-        // and delete it then we call function removePanic
-        if (player.grave[key].panic === false) {
-            // console.log('We delete panic!');
-            delete player.grave[key].panic;
-            removePanic(player);
-        }
-        delete player.hand[key];
-    }
-}
-
-function graveyardCheck(game, card, activeCard) {
-    const pActive = getActivePlayer(game);
-    const pInactive = getInActivePlayer(game);
-    // if active card is in item holder
-    if (Object.keys(pActive.item)[0] === card) {
-    // We move active card from item holder to graveyard
-        moveCardGraveyard(pActive, card, cardConst.ITEMCARD);
-    // if active player can chose opponent's card after act with turningPotion
-    // then we move card from opponent hand to his grave
-    } else if (pActive.turningHand === true) {
-        moveCardGraveyard(pInactive, card);
-    } else if (activeCard.disabled === false) {
-    // In other cases we move active card from hand to Graveyard
-        moveCardGraveyard(pActive, card);
-    }
-}
-
-function damagePlayer(player, points) {
-    // console.log('damagePlayer');
-    player.health.current -= points;
-}
-
-function healPlayer(player, points) {
-    // console.log('healPlayer');
-    if (player.health.current + points > player.health.maximum) {
-        player.health.current = player.health.maximum;
-    } else {
-        player.health.current += points;
-    }
-}
+const { moveToGraveyard } = require('./graveyard');
 
 function attackShield(player, itemKey, points, opponent) {
-    // console.log('attackShield');
     if (player.item[itemKey].healthCurrent > points) {
-    // console.log('item > points');
         player.item[itemKey].healthCurrent -= points;
     } else if (player.item[itemKey].healthCurrent === points) {
-    // console.log('item == points');
-        player.item[itemKey].healthCurrent = player.item[itemKey].health;
-
-        player.item[itemKey].fromOpponent === true
-            ? moveCardGraveyard(player, itemKey, cardConst.ITEMINACTIVE, opponent)
-            : moveCardGraveyard(player, itemKey, cardConst.ITEMCARD);
-    // console.log(player.grave[itemKey]);
+        moveToGraveyard(player, itemKey, opponent);
     } else {
-    // console.log('item < points', player.item[itemKey].healthCurrent, points);
         damagePlayer(player, points - player.item[itemKey].healthCurrent);
-        player.item[itemKey].healthCurrent = player.item[itemKey].health;
-        player.item[itemKey].fromOpponent === true
-            ? moveCardGraveyard(player, itemKey, cardConst.ITEMINACTIVE, opponent)
-            : moveCardGraveyard(player, itemKey, cardConst.ITEMCARD);
+        moveToGraveyard(player, itemKey, opponent);
     }
 }
 
@@ -119,7 +37,6 @@ function attackShield(player, itemKey, points, opponent) {
 // or in case of plateMail func reflect 1 point back to player
 // if active card has 1 attack points then no damage to both players
 function reflect(opponent, player, points) {
-    // console.log('We are in reflect function!', player);
     const itemCard = Object.values(opponent.item)[0].id;
     let damage;
     if (itemCard === cardConst.MAGICMIRRORCARD) {
@@ -144,12 +61,8 @@ function reflect(opponent, player, points) {
     ) {
         attackShield(player, Object.keys(player.item)[0], damage);
     }
-    if (player.health.current < 0) {
-        player.health.current = 0;
-    }
-    if (opponent.health.current < 0) {
-        opponent.health.current = 0;
-    }
+    setHealthToZero(player.health.current);
+    setHealthToZero(opponent.health.current);
 }
 
 // function to handle attack points deduction from opponent player health
@@ -182,115 +95,6 @@ function attackOpponent(player, opponent, points) {
     // console.log('Were in attack shield');
     // console.log(itemKey);
         attackShield(player, itemKey, points, opponent);
-    }
-}
-
-function pActiveIsTarget(game, activeCard, cardId) {
-    const pActive = getActivePlayer(game);
-    const pInactive = getInActivePlayer(game);
-    switch (activeCard.category) {
-    case cardConst.HEALCATEGORY:
-        healPlayer(pActive, activeCard.points);
-        // after player's act we change lastAction property of the game
-        lastActionChange(game, action.HEAL);
-        pActive.turningHand === true
-            ? moveCardGraveyard(pInactive, cardId)
-            : moveCardGraveyard(pActive, cardId);
-        break;
-    case cardConst.ATTACKCATEGORY:
-        break;
-    // if any mistake occurs during game process, player gets error message by default
-    default:
-        return new Error('You are under spell. Wait for redemption!');
-    }
-}
-
-// TODO: rename to Skull Lantern
-// this function runs when player attacks with the card Skull Lantern which category is attackItems
-// it accept players and checks all item category cards from both players item holders and hands
-// and move such cards to players grave yards
-function attackItems(players) {
-    players.forEach((p) => {
-        if (Object.keys(p.item).length !== 0) {
-            const itemCard = Object.values(p.item)[0];
-            // we reset item card's health points to initial points
-            itemCard.healthCurrent = itemCard.health;
-            // we move any item card to graveyard
-            moveCardGraveyard(p, Object.keys(p.item)[0], cardConst.ITEMCARD);
-        }
-        // we check whether each player hand is not empty
-        if (Object.keys(p.hand).length !== 0) {
-            // and for each card in hand with type item
-            for (const cardIndex in p.hand) {
-                if (p.hand[cardIndex].type === cardConst.ITEMCARD) {
-                    // we reset item card's points to initial points
-                    p.hand[cardIndex].healthCurrent = p.hand[cardIndex].health;
-                    // we move any item card to graveyard
-                    moveCardGraveyard(p, cardIndex);
-                }
-            }
-        }
-    });
-    return { players };
-}
-
-function pInactiveIsTarget(game, activeCard, cardId) {
-    const pActive = getActivePlayer(game);
-    const pInactive = getInActivePlayer(game);
-    // eslint-disable-next-line default-case
-    switch (activeCard.category) {
-    case cardConst.HEALCATEGORY:
-        break;
-    case cardConst.ATTACKCATEGORY:
-        attackOpponent(pInactive, pActive, activeCard.points);
-        pActive.turningHand === true
-            ? moveCardGraveyard(pInactive, cardId)
-            : moveCardGraveyard(pActive, cardId);
-        // after player's act we change lastAction property of the game
-        lastActionChange(game, action.ATACKOPPONENT);
-        break;
-    // if player attacks with card category holdCard we call disableCards function
-    // then move this attack card to gravyeard
-    case cardConst.HOLDCARDCATEGORY:
-        disableCards(pInactive);
-        pActive.turningHand === true
-            ? moveCardGraveyard(pInactive, cardId)
-            : moveCardGraveyard(pActive, cardId);
-        // after player's act we change lastAction property of the game
-        lastActionChange(game, action.CHAINS);
-        break;
-    // if player attacks with card category == attackItems, we call attack items function
-    // then move this attack card to gravyeard
-    case cardConst.ATTACKITEMSCATEGORY:
-        attackItems(game.players);
-        pActive.turningHand === true
-            ? moveCardGraveyard(pInactive, cardId)
-            : moveCardGraveyard(pActive, cardId);
-        // after player's act we change lastAction property of the game
-        lastActionChange(game, action.ATTACKITEMS);
-        break;
-    // if player attackes with clairvoyance card, we call showCards function
-    // then move this attack card to gravyeard
-    case cardConst.SHOWCARDSCATEGORY:
-        showCards(pInactive);
-        pActive.turningHand === true
-            ? moveCardGraveyard(pInactive, cardId)
-            : moveCardGraveyard(pActive, cardId);
-        // after player's act we change lastAction property of the game
-        lastActionChange(game, action.CLAIRVOYANCE);
-        break;
-    // if player attackes with turningPotion card, we call turningHand function
-    // then move this attack card to gravyeard
-    case cardConst.TURNINGCATEGORY:
-        turningHand(pActive, pInactive);
-        moveCardGraveyard(pActive, cardId);
-        // after player's act we change lastAction property of the game
-        lastActionChange(game, action.TURNINGPOTION);
-        break;
-
-    // if any mistake occurs during game process, player gets error message by default
-    default:
-        return new Error('You are under spell. Wait for redemption!');
     }
 }
 
@@ -333,7 +137,7 @@ function moveItem(game, player, key, opponent) {
 function changeMoveCounter(pActive, card, pInactive) {
     if (pActive.hand[card] === undefined && pActive.turningHand === false) {
         pActive.moveCounter += 1;
-        const noClairvoyance = (player) => setTimeout(() => deleteCardsShown(player), 1000);
+        const noClairvoyance = (player) => setTimeout(() => deleteCardsShownProperty(player), 1000);
         noClairvoyance(pInactive);
         noClairvoyance(pActive);
     }
@@ -381,17 +185,10 @@ function forestMushroom(game, opponent) {
 }
 
 module.exports = {
-    getActivePlayer,
-    getInActivePlayer,
-    lastActionChange,
-    moveCardGraveyard,
-    graveyardCheck,
     attackShield,
-    pActiveIsTarget,
-    attackItems,
-    pInactiveIsTarget,
     moveItem,
     changeMoveCounter,
     changeTurn,
     forestMushroom,
+    attackOpponent,
 };
